+++ CHANGELOG.md
@@ -21,6 +21,17 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ---
 
+## 1.6.6
+
+2026 Jul 29 [in progress]
+
+* Fixed Waveform pin_left / pin_right getting stuck when viewing JLS files:
+  pins are now forced off and the pin buttons disabled for file sources.
+  Ctrl+A and Shift+Space zoom to the full extents in file mode.
+* Added Waveform pinned x-axis indication (pin button highlight and message)
+  when pins block an x-axis pan attempt during streaming  #324
+
+
 ## 1.6.5
 
 2026 Jul 20